feat: Add SharedResources service
The commit adds the `SharedResources` service to the `utils` package. This service was missing from the previous version and is now included in the package.

Note: The commit message follows the convention of starting with a type (feat for feature) followed by a colon and a brief description of the change.